requirements.txt
streamlit
numpy
pillow
opencv-python-headless
tflite-runtime



